requirements.txt
aiohttp==3.7.4.post0
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cabocha @ file:///Users/yuki/Documents/GitHub/line-bot-sdk-python/mecab-ipadic-neologd/cabocha-master/cabocha
cabocha-python @ file:///Users/yuki/Documents/GitHub/line-bot-sdk-python/mecab-ipadic-neologd/cabocha-master/python
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cycler==0.10.0
discord.py==1.7.3
fitbit==0.3.1
Flask==2.0.1
idna==2.10
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mecab-python3==1.0.4
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
Pillow==8.2.0
platformdirs==2.3.0
pycodestyle==2.7.0
pycparser==2.20
pylint==2.10.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
slackbot==1.0.0
slacker==0.14.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.5
websocket-client==0.44.0
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
